+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 ### BREAKING CHANGES
 
+## [2.1.0] - 2018-08-17
+### Added
+- Add logLevel option.
+- Add describe property to suites. Log the description when suite starts.
+
 ## [2.0.0] - 2018-08-12
 ### Added
 - Add support for "include" yaml feature, that allows to split configuration into many partial files.